+++ starlight-core/src/main/java/org/atcraftmc/starlight/core/platform/BukkitUtil.java
@@ -26,8 +26,10 @@ import org.bukkit.util.Vector;
 
 import java.lang.reflect.Method;
 import java.text.DecimalFormat;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Random;
+import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Future;
 import java.util.function.Consumer;