+++ starlight-extensions/starlight-ext-ai/src/main/java/org/atcraftmc/starlight/SLAiPackage.java
@@ -3,6 +3,7 @@ package org.atcraftmc.starlight;
 import me.gb2022.gluon.pack.ApplicationPackageProvider;
 import me.gb2022.gluon.pack.ContentBuilder;
 import org.atcraftmc.starlight.ai.AIChatService;
+import org.atcraftmc.starlight.ai.AICommandChat;
 import org.atcraftmc.starlight.framework.PluginPackageAttachment;
 import org.atcraftmc.starlight.framework.pack.MultiPackageProvider;
 import org.atcraftmc.starlight.framework.pack.SLPackageProvider;
@@ -14,5 +15,9 @@ public final class SLAiPackage extends MultiPackageProvider {
         var i = b.getAttachment(PluginPackageAttachment.class);
 
         b.service(AIChatService.class);
+
+        b.module(AICommandChat.class);
+
+        i.language("starlight-ai", "zh_cn");
     }
 }