+++ starlight-extensions/starlight-ext-ai/src/main/java/org/atcraftmc/starlight/ai/chat/ChatRequest.java
@@ -1,4 +1,4 @@
 package org.atcraftmc.starlight.ai.chat;
 
-public record ChatRequest(String systemPrompt, String userInput) {
+public record ChatRequest(String systemPrompt, String userInput,String username,String contextId) {
 }