+++ starlight-game/src/main/java/org/atcraftmc/starlight/tweaks/PortableShulkerBox.java
@@ -12,6 +12,8 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.block.BlockPlaceEvent;
+import org.bukkit.event.entity.PlayerDeathEvent;
+import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.inventory.InventoryCloseEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
@@ -22,60 +24,89 @@ import org.bukkit.inventory.meta.BlockStateMeta;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
 
 @ApplicationModule(id = "portable-shulker-box", version = "1.2", description = "Opens shulker boxes directly from the inventory")
 @AutoRegister(Registrations.SERVER_EVENT)
 public final class PortableShulkerBox extends BukkitAbstractModule {
-    private final Map<String, Session> sessions = new HashMap<>();
+    private final Map<UUID,Session> sessions = new ConcurrentHashMap<>();
 
     @Override
     public void disable() {
-        for (Player p : Bukkit.getOnlinePlayers()) {
-            this.close(p.getName());
+        for (var p : Bukkit.getOnlinePlayers()) {
+            this.close(p);
         }
     }
+
+    public void open(Player player) {
+        var inv = player.getInventory();
+
+        if (!inv.getItemInMainHand().getType().getKey().getKey().endsWith("shulker_box")) {
+            return;
+        }
+
+        var session = new Session(inv.getItemInMainHand(), player);
+        this.sessions.put(player.getUniqueId(), session);
+    }
+
+    public void close(Player player){
+        this.sync(player);
+        this.sessions.remove(player.getUniqueId());
+    }
+
+    public void sync(Player player) {
+        var uuid = player.getUniqueId();
+
+        if (!this.sessions.containsKey(player.getUniqueId())) {
+            return;
+        }
+
+        this.sessions.get(uuid).sync();
+    }
+
 
     @EventHandler
     public void onPlayerInteract(PlayerInteractEvent event) {
         if (event.getAction() != Action.RIGHT_CLICK_AIR) {
             return;
         }
-        PlayerInventory inv = event.getPlayer().getInventory();
 
-        if (!inv.getItemInMainHand().getType().getKey().getKey().endsWith("shulker_box")) {
-            return;
-        }
-
-        QLib.task().entity(event.getPlayer()).run(() -> {
-            Session session = new Session(inv.getItemInMainHand(), event.getPlayer());
-            this.sessions.put(event.getPlayer().getName(), session);
-        });
+        this.open(event.getPlayer());
     }
 
-    public void close(String player) {
-        if (!this.sessions.containsKey(player)) {
+    @EventHandler
+    public void onInventoryClick(InventoryClickEvent event) {
+        if(!(event.getWhoClicked() instanceof Player player)) {
             return;
         }
-        Entity entity = Bukkit.getPlayerExact(player);
-        QLib.task().entity(entity).run(() -> {
-            this.sessions.get(player).close();
-            this.sessions.remove(player);
-        });
+
+        this.sync(player);
     }
 
     @EventHandler
     public void onInventoryClose(InventoryCloseEvent event) {
-        this.close(event.getPlayer().getName());
+        if(!(event.getPlayer() instanceof Player player)) {
+            return;
+        }
+
+        this.close(player);
     }
 
     @EventHandler
     public void onPlayerQuit(PlayerQuitEvent event) {
-        this.close(event.getPlayer().getName());
+        this.close(event.getPlayer());
     }
 
     @EventHandler
+    public void onPlayerDeath(PlayerDeathEvent event) {
+        this.close(event.getEntity());
+    }
+
+
+    @EventHandler
     public void onBlockPlace(BlockPlaceEvent event) {
-        if (this.sessions.containsKey(event.getPlayer().getName())) {
+        if (this.sessions.containsKey(event.getPlayer().getUniqueId())) {
             event.setCancelled(true);
         }
     }
@@ -94,7 +125,7 @@ public final class PortableShulkerBox extends BukkitAbstractModule {
             target.openInventory(inventory);
         }
 
-        public void close() {
+        public void sync() {
             this.meta.setBlockState(state);
             this.item.setItemMeta(this.meta);
         }